Let the ssl context take the ownership
This makes it easier to handle the SSL TCP stack.

When the SSL context has only a reference, the
original (delegate) context must effectively be
'static.

On the other hand, the SSL stack can take over
the plain TcpStack, offer a reference to the
delegate, and also implement traits the delegate
has.

Also see:
  https://stackoverflow.com/a/30824188